requirement.txt
colorama==0.4.4
commonmark==0.9.1
numpy==1.22.2
Pillow==9.0.1
Pygments==2.11.2
rich==11.2.0
